library refactor average waveforms
