+++ src/libraries/AmplitudeScaleToolbarEntries/AmplitudeScaleToolbarEntries.tsx
