+++ src/libraries/AmplitudeScaleToolbarEntries/index.ts
@@ -1,0 +1,1 @@
+export {default as AmplitudeScaleToolbarEntries} from './AmplitudeScaleToolbarEntries'
