+++ src/libraries/AverageWaveforms/AverageWaveformPlot.tsx
