+++ src/libraries/AverageWaveforms/AverageWaveformsViewData.ts
