+++ src/libraries/AverageWaveforms/WaveformWidget/WaveformPlot.tsx
