+++ src/libraries/AverageWaveforms/WaveformWidget/WaveformWidget.tsx
