+++ src/libraries/AverageWaveforms/WaveformWidget/sharedDrawnComponents/ElectrodeGeometry.tsx
