+++ src/libraries/AverageWaveforms/WaveformWidget/sharedDrawnComponents/ElectrodeGeometrySvg.tsx
