+++ src/libraries/AverageWaveforms/WaveformWidget/sharedDrawnComponents/electrodeGeometryLayout.ts
