+++ src/libraries/AverageWaveforms/WaveformWidget/sharedDrawnComponents/electrodeGeometryPainting.ts
