+++ src/libraries/AverageWaveforms/WaveformWidget/sharedDrawnComponents/electrodeGeometryStateManagement.tsx
