+++ src/libraries/AverageWaveforms/WaveformWidget/sharedDrawnComponents/utility.ts
