+++ src/libraries/AverageWaveforms/WaveformWidget/waveformLogic.ts
