+++ src/libraries/AverageWaveforms/index.ts
@@ -1,0 +1,7 @@
+export { default as AverageWaveformsView } from './AverageWaveformsView'
+export { isAverageWaveformsViewData } from './AverageWaveformsViewData'
+export type { AverageWaveformsViewData } from './AverageWaveformsViewData'
+export { default as ElectrodeGeometry } from './WaveformWidget/sharedDrawnComponents/ElectrodeGeometry'
+export { computeElectrodeLocations, computeElectrodesFromIdsAndLocations } from './WaveformWidget/sharedDrawnComponents/electrodeGeometryLayout'
+export { defaultColors } from './WaveformWidget/sharedDrawnComponents/electrodeGeometryPainting'
+export type { ElectrodeColors } from './WaveformWidget/sharedDrawnComponents/electrodeGeometryPainting'
