+++ src/views/RawTraces/RawTracesComponent.tsx
@@ -1,9 +1,9 @@
+import { AmplitudeScaleToolbarEntries } from 'libraries/AmplitudeScaleToolbarEntries'
 import { useRecordingSelectionTimeInitialization, useTimeRange } from 'libraries/RecordingSelectionContext'
 import { colorForUnitId } from 'libraries/UnitColors'
 import { matrix, multiply } from 'mathjs'
 import { FunctionComponent, useCallback, useEffect, useMemo, useState } from 'react'
 import { TimeseriesLayoutOpts } from 'View'
-import AmplitudeScaleToolbarEntries from 'views/common/AmplitudeScaleToolbarEntries'
 import { DefaultToolbarWidth } from 'views/common/TimeWidgetToolbarEntries'
 import useFetchCache from 'views/PositionPdfPlot/useFetchCache'
 import TimeScrollView, { use1dTimeToPixelMatrix, usePanelDimensions, usePixelsPerSecond, useTimeseriesMargins } from '../RasterPlot/TimeScrollView/TimeScrollView'
