+++ src/views/SpikeAmplitudes/SpikeAmplitudesView.tsx
@@ -1,11 +1,11 @@
+import { AmplitudeScaleToolbarEntries } from 'libraries/AmplitudeScaleToolbarEntries'
 import { useRecordingSelectionTimeInitialization, useTimeRange } from 'libraries/RecordingSelectionContext'
 import { Splitter } from 'libraries/Splitter'
 import { colorForUnitId } from 'libraries/UnitColors'
 import { idToNum } from 'libraries/UnitSelectionContext'
 import { matrix, multiply } from 'mathjs'
-import React, { FunctionComponent, useMemo, useState } from 'react'
+import { FunctionComponent, useMemo, useState } from 'react'
 import { TimeseriesLayoutOpts } from 'View'
-import AmplitudeScaleToolbarEntries from 'views/common/AmplitudeScaleToolbarEntries'
 import LockableSelectUnitsWidget from 'views/common/SelectUnitsWidget/LockableSelectUnitsWidget'
 import useLocalSelectedUnitIds from 'views/common/SelectUnitsWidget/useLocalSelectedUnitIds'
 import useYAxisTicks from 'views/common/TimeScrollView/YAxisTicks'
